Fix two critical bugs
Bug 1: Piano roll UI components disappearing on re-entry
- StatusLine, Toolbar, Inspector were hidden but not re-shown
- Added setVisible(true) when components already exist
- Now properly shows all UI when reopening piano roll

Bug 2: Recording velocity always minimum value
- MidiRecorder was using getFloatVelocity() (0.0-1.0 range)
- Converting float to int resulted in 0 or 1
- Changed to getVelocity() which returns 0-127
- Updated ActiveNote.velocity from float to int
- Updated handleNoteOn signature to accept int velocity

+++ src/core/audio/MidiRecorder.h
@@ -79,7 +79,7 @@ private:
     struct ActiveNote
     {
         int pitch;
-        float velocity;
+        int velocity;  // Changed from float to int (0-127)
         int64_t startTick;
     };
     std::map<int, ActiveNote> activeNotes;  // key = MIDI note number
@@ -89,7 +89,7 @@ private:
     
     // Helper methods
     int64_t quantizeTick(int64_t tick) const;
-    void handleNoteOn(int noteNumber, float velocity, int64_t tick);
+    void handleNoteOn(int noteNumber, int velocity, int64_t tick);  // Changed from float to int
     void handleNoteOff(int noteNumber, int64_t tick);
     void clearRegion(int64_t startTick, int64_t endTick);
